Move Capability to api subproject
