+++ impl/src/main/scala/quasar/impl/datasource/local/LocalDatasourceModule.scala
@@ -17,9 +17,9 @@
 package quasar.impl.datasource.local
 
 import quasar.Disposable
-import quasar.api.datasource.{DatasourceError, DatasourceType}, DatasourceError._
+import quasar.api.datasource.{Capability, DatasourceError, DatasourceType}, DatasourceError._
 import quasar.concurrent.BlockingContext
-import quasar.connector.{Capability, LightweightDatasourceModule, MonadResourceErr}
+import quasar.connector.{LightweightDatasourceModule, MonadResourceErr}
 
 import scala.concurrent.ExecutionContext
