+++ impl/src/main/scala/quasar/impl/datasource/local/LocalParsedDatasourceModule.scala
@@ -17,10 +17,10 @@
 package quasar.impl.datasource.local
 
 import quasar.Disposable
-import quasar.api.datasource.{DatasourceError, DatasourceType}, DatasourceError._
+import quasar.api.datasource.{Capability, DatasourceError, DatasourceType}, DatasourceError._
 import quasar.common.data.RValue
 import quasar.concurrent.BlockingContext
-import quasar.connector.{Capability, LightweightDatasourceModule, MonadResourceErr}
+import quasar.connector.{LightweightDatasourceModule, MonadResourceErr}
 
 import scala.concurrent.ExecutionContext
